+++ src/components/Gantt.vue
@@ -68,6 +68,7 @@ defineSlots<{
   bars?: (props: { tasks: unknown }) => unknown
   grid?: (props: { columns: unknown; rows: unknown }) => unknown
   conflicts?: (props: { conflicts: unknown }) => unknown
+  slack?: (props: { slack: unknown }) => unknown
   dependencies?: (props: { tasks: unknown }) => unknown
   today?: (props: { today: unknown; dateToX: unknown }) => unknown
   'body-extra'?: (props: { contentWidth: number; contentHeight: number }) => unknown
@@ -142,6 +143,9 @@ defineExpose({
       </template>
       <template v-if="$slots.conflicts" #conflicts="slotProps">
         <slot name="conflicts" v-bind="slotProps" />
+      </template>
+      <template v-if="$slots.slack" #slack="slotProps">
+        <slot name="slack" v-bind="slotProps" />
       </template>
       <template v-if="$slots.dependencies" #dependencies="slotProps">
         <slot name="dependencies" v-bind="slotProps" />