+++ src/components/GanttView.vue
@@ -7,6 +7,7 @@ import GanttDependencies from './GanttDependencies.vue'
 import GanttGrid from './GanttGrid.vue'
 import GanttGroupBar from './GanttGroupBar.vue'
 import GanttMilestone from './GanttMilestone.vue'
+import GanttSlack from './GanttSlack.vue'
 import GanttTask from './GanttTask.vue'
 import GanttTaskList from './GanttTaskList.vue'
 import GanttTimeline from './GanttTimeline.vue'
@@ -28,6 +29,7 @@ const {
   tasks,
   config,
   conflicts,
+  slack,
   visibleColumnsFor,
   dateToX,
   contentWidth,
@@ -126,6 +128,10 @@ const scrollStyle = computed(() => {
         </slot>
         <slot name="conflicts" :conflicts="conflicts">
           <GanttConflicts v-if="config.overlap === 'conflict'" />
+        </slot>
+
+        <slot name="slack" :slack="slack">
+          <GanttSlack v-if="config.slack" />
         </slot>
 
         <slot name="dependencies" :tasks="tasks">